alignment: avoid truncation on 32 bit systems

+++ src/lib/jp2/util/MemManager.cpp
@@ -68,7 +68,7 @@ namespace grk {
 const uint32_t grk_alignment = 32;
 
 uint32_t grk_make_aligned_width(uint32_t width){
-	return (uint32_t)((((size_t)width + grk_alignment - 1)/grk_alignment) * grk_alignment);
+	return (uint32_t)((((uint64_t)width + grk_alignment - 1)/grk_alignment) * grk_alignment);
 }
 
 static inline void* grk_aligned_alloc_n(size_t alignment, size_t size) {
